backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
setuptools
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
django-clearcache==1.2.1
django-modeladmin-reorder==0.3.1
